payu_service/requirements.txt
quixstreams
jsonschema
python-dotenv